+++ packages/react-reconciler/src/ReactFiberFlags.js
@@ -42,6 +42,7 @@ export const StoreConsistency = /*             */ 0b0000000000000000100000000000
 // It's OK to reuse these bits because these flags are mutually exclusive for
 // different fiber types. We should really be doing this for as many flags as
 // possible, because we're about to run out of bits.
+export const Hydrate = Callback;
 export const ScheduleRetry = StoreConsistency;
 export const ShouldSuspendCommit = Visibility;
 export const ViewTransitionNamedMount = ShouldSuspendCommit;